Add files via upload
two base files of what will hopefully be a .exe eventually. tktrials.py is so named because it was my first foray into creating a GUI and I was doing it with tkinter. It is the main function in the group- EPICscrape was created prior and originally meant to be run from the command line now it serves a a workhorse for the grabbing and manipulating of scraped data.- DM 2018